chore: tetris attack can be float

+++ combo.h
@@ -9,13 +9,13 @@
 
 #include <functional>
 
-using ComboFunc = std :: function <int(
+using ComboFunc = std :: function <float(
 	int clear, Spin spin, int combo, int b2b, bool pc
 )>;
 
 namespace KeroshiCombo {
-	inline int stage1(int clear, Spin spin, int combo, int b2b, bool pc) {
-		constexpr int table[5] = {
+	inline float stage1(int clear, Spin spin, int combo, int b2b, bool pc) {
+		constexpr float table[5] = {
 			0, 1, 3, 6, 12
 		};
 		return table[clear];
